+++ Assets/Scripts/Interfaces/IHitDetectable.cs
@@ -4,7 +4,7 @@ namespace Interfaces
 {
     public interface IHitDetectable
     {
-        public double OnHit(Armor armor);
+        public float OnHit(Armor armor);
         public void OnHitTaken(Character attacker);
     }
 }
