+++ Assets/Scripts/Interfaces/IMagicalRpgObject.cs
@@ -2,7 +2,7 @@
 {
     public interface IMagicalRpgObject
     {
-        double MagicFactor { set; } // either damage or resistance modifier
+        float MagicFactor { set; } // either damage or resistance modifier
         string ElementalFactor { set; }
     }
 }
